# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartConvert API - file conversion, compression, adaptive compression and benchmarking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ConversionType = typeof ConversionType[keyof typeof ConversionType];


export const ConversionType = {
  'docx-to-pdf': 'docx-to-pdf',
  'pdf-to-docx': 'pdf-to-docx',
  'pdf-to-txt': 'pdf-to-txt',
  'pdf-to-markdown': 'pdf-to-markdown',
  'images-to-pdf': 'images-to-pdf',
  'pdf-to-images': 'pdf-to-images',
} as const;

export interface ConversionInput {
  files: Blob[];
  conversionType: ConversionType;
}

export interface ConversionOutputFile {
  id: string;
  fileName: string;
  fileSize: number;
  downloadUrl: string;
}

export interface ConversionResult {
  id: string;
  conversionType: ConversionType;
  sourceFileNames: string[];
  outputs: ConversionOutputFile[];
  /** @nullable */
  warning?: string | null;
  createdAt: string;
}

export type CompressionAlgorithm = typeof CompressionAlgorithm[keyof typeof CompressionAlgorithm];


export const CompressionAlgorithm = {
  zip: 'zip',
  gzip: 'gzip',
  bzip2: 'bzip2',
  lzma: 'lzma',
  brotli: 'brotli',
  zstd: 'zstd',
  lz4: 'lz4',
  snappy: 'snappy',
} as const;

export interface CompressionInput {
  files: Blob[];
  algorithm: CompressionAlgorithm;
}

export interface CompressionResult {
  id: string;
  fileName: string;
  algorithm: CompressionAlgorithm;
  originalSize: number;
  compressedSize: number;
  compressionRatio: number;
  compressionTimeMs: number;
  decompressionTimeMs: number;
  compressionSpeedMBps: number;
  downloadUrl: string;
  createdAt: string;
}

export interface AdaptiveCompressionInput {
  files: Blob[];
}

export interface FileAnalysis {
  mimeType: string;
  isText: boolean;
  entropy: number;
  alreadyCompressed: boolean;
  repetitionScore: number;
  /** @nullable */
  avgLineLength: number | null;
}

export interface AdaptiveCompressionResult {
  id: string;
  fileName: string;
  algorithm: CompressionAlgorithm;
  originalSize: number;
  compressedSize: number;
  compressionRatio: number;
  compressionTimeMs: number;
  decompressionTimeMs: number;
  compressionSpeedMBps: number;
  downloadUrl: string;
  analysis: FileAnalysis;
  reasons: string[];
  createdAt: string;
}

export interface BenchmarkInput {
  files: Blob[];
  /** Comma-separated list of CompressionAlgorithm values */
  algorithms: string;
}

export interface BenchmarkEntry {
  algorithm: CompressionAlgorithm;
  compressedSize: number;
  compressionRatio: number;
  compressionTimeMs: number;
  decompressionTimeMs: number;
  /** @nullable */
  memoryUsageKb: number | null;
}

export interface BenchmarkResult {
  id: string;
  fileName: string;
  originalSize: number;
  entries: BenchmarkEntry[];
  bestAlgorithm: CompressionAlgorithm;
  createdAt: string;
}

export type HistoryEntryType = typeof HistoryEntryType[keyof typeof HistoryEntryType];


export const HistoryEntryType = {
  convert: 'convert',
  compress: 'compress',
  adaptive: 'adaptive',
  benchmark: 'benchmark',
} as const;

export interface HistoryEntry {
  id: string;
  type: HistoryEntryType;
  summary: string;
  /** @nullable */
  detail?: string | null;
  createdAt: string;
}