# lib/api-zod/src/generated/types/adaptiveCompressionResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartConvert API - file conversion, compression, adaptive compression and benchmarking
 * OpenAPI spec version: 0.1.0
 */
import type { CompressionAlgorithm } from './compressionAlgorithm';
import type { FileAnalysis } from './fileAnalysis';

export interface AdaptiveCompressionResult {
  id: string;
  fileName: string;
  algorithm: CompressionAlgorithm;
  originalSize: number;
  compressedSize: number;
  compressionRatio: number;
  compressionTimeMs: number;
  decompressionTimeMs: number;
  compressionSpeedMBps: number;
  downloadUrl: string;
  analysis: FileAnalysis;
  reasons: string[];
  createdAt: Date;
}