# lib/api-zod/src/generated/types/benchmarkResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartConvert API - file conversion, compression, adaptive compression and benchmarking
 * OpenAPI spec version: 0.1.0
 */
import type { BenchmarkEntry } from './benchmarkEntry';
import type { CompressionAlgorithm } from './compressionAlgorithm';

export interface BenchmarkResult {
  id: string;
  fileName: string;
  originalSize: number;
  entries: BenchmarkEntry[];
  bestAlgorithm: CompressionAlgorithm;
  createdAt: Date;
}