# lib/api-zod/src/generated/types/conversionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartConvert API - file conversion, compression, adaptive compression and benchmarking
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionType } from './conversionType';

export interface ConversionInput {
  files: Blob[];
  conversionType: ConversionType;
}