# lib/api-zod/src/generated/types/conversionOutputFile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartConvert API - file conversion, compression, adaptive compression and benchmarking
 * OpenAPI spec version: 0.1.0
 */

export interface ConversionOutputFile {
  id: string;
  fileName: string;
  fileSize: number;
  downloadUrl: string;
}