# lib/api-zod/src/generated/types/conversionResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartConvert API - file conversion, compression, adaptive compression and benchmarking
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionOutputFile } from './conversionOutputFile';
import type { ConversionType } from './conversionType';

export interface ConversionResult {
  id: string;
  conversionType: ConversionType;
  sourceFileNames: string[];
  outputs: ConversionOutputFile[];
  /** @nullable */
  warning?: string | null;
  createdAt: Date;
}