# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartConvert API - file conversion, compression, adaptive compression and benchmarking
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntryType } from './historyEntryType';

export interface HistoryEntry {
  id: string;
  type: HistoryEntryType;
  summary: string;
  /** @nullable */
  detail?: string | null;
  createdAt: Date;
}